requirements.txt
flask
langchain
langchain-mistralai
python-dotenv
mistralai
gunicorn